Add unit tests for various services and components
- Implement tests for RedisPublicRateLimitService to validate rate limiting behavior.
- Create AnalyticsReadService tests to ensure analytics retrieval for owned apps.
- Add AppKeyService tests to verify app key validation and rejection of overlong keys.
- Develop AppService tests to check app creation and deletion logic.
- Introduce AuthService tests to reject overlong passphrases during login.
- Implement HealthService tests to validate application and service health checks.
- Add RateLimitService tests to ensure proper rate limiting and analytics recording.
- Create Redis-based rate limiter tests to validate Redis interactions and response mapping.
- Implement RuleService tests to verify rule creation, deletion, and updates.
- Add analytics persistence tests to ensure proper snapshot handling and cleanup.
- Introduce RedisAnalyticsSnapshotReader tests to validate reading from Redis.
- Add logback configuration for testing to control logging output.

+++ src/main/java/com/desertrider/throttlr/dto/response/AppResponse.java
@@ -1,5 +1,5 @@
 package com.desertrider.throttlr.dto.response;
 
-public record AppResponse(String appId, String name, int ruleCount, long createdAt) {
+public record AppResponse(String appId, String name, long ruleCount, long createdAt) {
 
 }
